+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix it should update avatar without cache in the next call (issue 958)
 - Fix it should display name from phone book in calls (issue 960)
 - Fix android it should have notification sound after a missed group call answer simultaneously (issue 974)
+- Fix it should show UC notification in call manage screen (issue 982)
 
 #### 2.16.2